C++: remove C-specific code to compile with g++
- Replace designated initializers with init functions
- Rename or/xor/and to nemuor/nemuxor/nemuand
- Replace (void *) with (type *), and explicitly convert types
- RV64 System/User can be compiled; RV64 System has been tested.

+++ src/engine/interpreter/fp.c
@@ -9,7 +9,7 @@ static inline rtlreg_t unbox(rtlreg_t r) {
 }
 
 static inline float32_t rtlToF32(rtlreg_t r) {
-  float32_t f = { .v = unbox(r) };
+  float32_t f = { .v = (uint32_t)unbox(r) };
   return f;
 }
